Revival of Project
Resurrected the project from the dead.

New method of finding Steam's drive and directory as well as the Among Us directories.

Switched from WebClient to HttpClient. (Removal of nWeb.cs)

Removed LNK2Path as this method is no longer used to find Steam's drive and directory.

Added global.json to specify to compile with preview 6 of .net6.0 and not anything above or below. This can be removed however and still compiled, but .net6.0 preview 7 gives interesting Linker errors that seem to be unreasonable while the rest do not.